+++ src/citra_qt/configuration/groups/config_shortcuts.cpp
@@ -7,7 +7,7 @@
 #include "citra_qt/uisettings.h"
 
 // clang-format off
-const std::array<UISettings::Shortcut, 35> default_hotkeys{
+const std::array<UISettings::Shortcut, 36> default_hotkeys{
     {{QStringLiteral("2x Native Internal Resolution"),     QStringLiteral("Main Window"), {QStringLiteral("Alt+2"), Qt::ApplicationShortcut}},
      {QStringLiteral("3x Native Internal Resolution"),     QStringLiteral("Main Window"), {QStringLiteral("Alt+3"), Qt::ApplicationShortcut}},
      {QStringLiteral("4x Native Internal Resolution"),     QStringLiteral("Main Window"), {QStringLiteral("Alt+4"), Qt::ApplicationShortcut}},
@@ -42,7 +42,8 @@ const std::array<UISettings::Shortcut, 35> default_hotkeys{
      {QStringLiteral("Toggle Texture Dumping"),            QStringLiteral("Main Window"), {QStringLiteral("Ctrl+D"), Qt::ApplicationShortcut}},
      {QStringLiteral("Toggle Hardware Shader"),            QStringLiteral("Main Window"), {QStringLiteral("Ctrl+H"), Qt::ApplicationShortcut}},
      {QStringLiteral("Toggle Custom Layout"),              QStringLiteral("Main Window"), {QStringLiteral("Ctrl+C"), Qt::ApplicationShortcut}},
-     {QStringLiteral("Toggle Custom Screen Refresh Rate"), QStringLiteral("Main Window"), {QStringLiteral("Ctrl+R"), Qt::ApplicationShortcut}}}};
+     {QStringLiteral("Toggle Custom Screen Refresh Rate"), QStringLiteral("Main Window"), {QStringLiteral("Ctrl+R"), Qt::ApplicationShortcut}},
+     {QStringLiteral("Change Custom Screen Refresh Rate"), QStringLiteral("Main Window"), {QStringLiteral("Ctrl+Alt+R"), Qt::ApplicationShortcut}}}};
 // clang-format on
 
 void Config::ReadShortcutsValues() {